fix(trusty-search): set skip_kg on create/re-register path + regression tests (refs #2984)
Code-critic review of PR #2988 found the create/re-register door
(POST /indexes) still loaded a persisted symbol graph for a skip_kg=true
request against an existing id, because create_index_handler built
init_entry via ..Default::default() (skip_kg=false) before the real
skip_kg value was computed further down the handler. skip_kg is now
computed before init_entry is constructed and set on the entry so
build_indexer_from_entry honors it on this call, not just after a
daemon restart.

Also adds the two missing regression tests the review called for:
- persistence_loader::tests::skip_kg_true_entry_never_loads_persisted_graph_via_build_indexer_from_entry
  exercises the real build_indexer_from_entry construction path end to
  end (build with skip_kg=false, persist a real graph, reopen the same
  id/root with skip_kg=true, assert the graph stays empty).
- core::indexer::tests::branch_and_corpus::skip_kg_true_legacy_json_restore_skips_graph_rebuild
  covers the previously-untested load_chunks_from_disk guard on the
  legacy chunks.json warm-boot path.

persistence_loader.rs was split into a persistence_loader/ directory
module (mod.rs + tests.rs) to stay under the 500-SLOC production cap
after the new test landed (issue #1195 convention, mirrors
core::indexer::tests).
