+++ crates/trusty-search/CHANGELOG.md
@@ -49,7 +49,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - migrate off archived/unsound `serde_yml` (GHSA-hhw4-xg65-fp2x) and its transitive `libyml` (GHSA-gfxp-f68g-8x78) onto the workspace-blessed `serde_yaml` 0.9, closing Dependabot alerts #42/#43 (closes [#2991](https://github.com/bobmatnyc/trusty-tools/issues/2991))
-- warm-boot now honors `skip_kg`: a `skip_kg=true` index previously loaded — or, on a first/never-persisted graph, fully rebuilt — its symbol graph into memory on every daemon restart regardless of the flag, because `IndexHandle::skip_kg` never reached the `CodeIndexer` that owns the actual load/rebuild call. `CodeIndexer` now carries its own `skip_kg` mirror (set by `build_indexer_from_entry` from the persisted entry), and `load_or_rebuild_symbol_graph` checks it before touching the corpus at all — neither the persisted-graph load nor the from-chunks rebuild runs for a skip_kg index, so the ~50-100 MB/index petgraph heap cost the flag was meant to avoid now actually survives a restart. Pure soft-off: on-disk KG tables are untouched; query-time degradation (`search_capabilities` excludes `kg`, `call_chain`/`search_kg` 503 `kg_unavailable`) is unchanged (Phase 0 of #2984)
+- warm-boot now honors `skip_kg`: a `skip_kg=true` index previously loaded — or, on a first/never-persisted graph, fully rebuilt — its symbol graph into memory on every daemon restart regardless of the flag, because `IndexHandle::skip_kg` never reached the `CodeIndexer` that owns the actual load/rebuild call. `CodeIndexer` now carries its own `skip_kg` mirror (set by `build_indexer_from_entry` from the persisted entry), and `load_or_rebuild_symbol_graph` checks it before touching the corpus at all — neither the persisted-graph load nor the from-chunks rebuild runs for a skip_kg index, so the ~50-100 MB/index petgraph heap cost the flag was meant to avoid now actually survives a restart. Pure soft-off: on-disk KG tables are untouched; query-time degradation (`search_capabilities` excludes `kg`, `call_chain`/`search_kg` 503 `kg_unavailable`) is unchanged. A code-critic follow-up review of the same PR caught a second door to the identical bug: `create_index_handler` (`POST /indexes`, the re-register path) computed the real `skip_kg` value AFTER already building the indexer via `Default::default()` (`skip_kg = false`), so re-registering an existing id with `skip_kg=true` still loaded its persisted graph during that same call. `skip_kg` is now computed before `init_entry` is constructed and set on the entry so `build_indexer_from_entry` honors it on the create/re-register door the same way it does on daemon-restart warm-boot (Phase 0 of #2984)
 - `embedder_supervisor::SupervisorConfig`: new `wedge_reset_secs` field, read from `TRUSTY_EMBEDDERD_WEDGE_RESET_SECS` (default 300s) in `from_env()` and threaded into both `trusty_common::embedder_client::SupervisorConfig` construction sites (`do_spawn`'s production spawn path and `into_common_for_tests`) — the env var previously had no effect on the running daemon because both sites built the common config with `..SupervisorConfig::default()`, silently discarding any operator override. A wedged-but-alive sidecar's restart-storm escalation counter (trusty-common #1448/#1450) is now genuinely tunable without recompiling ([#1448](https://github.com/bobmatnyc/trusty-tools/issues/1448), [#1450](https://github.com/bobmatnyc/trusty-tools/issues/1450))
 - LaunchAgent plist now sets an explicit 8192 fd_limit (matching the trusty-memory convention) so fresh installs warm-boot large index fleets without hitting launchd's default 256-fd ceiling; existing installs must re-run `trusty-search service install` to regenerate the plist (closes #2947)
 - shutdown-flush no longer truncates HNSW on timeout: `UsearchStore::save` now serializes concurrent savers of the same store, the graceful-shutdown flush deadline scales with each index's on-disk snapshot size (was a flat 10s) and flushes indexes with bounded concurrency instead of strictly sequentially, and a truncated/corrupt on-disk HNSW snapshot is now detected on load (via a pre-`view()` size floor plus a post-view zero-vector-vs-populated-sidecar guard) so `/health` no longer reports semantic search ready when it silently fell back to BM25-only (closes #2922) ([#2930](https://github.com/bobmatnyc/trusty-tools/pull/2930)) ([`f212b2e`](https://github.com/bobmatnyc/trusty-tools/commit/f212b2e1a1e8c74e29a6cf407023599ae3cb53bf))
