--- crates/trusty-search/src/service/persistence_loader.rs
@@ -1,746 +1,0 @@
-//! Shared helper that builds a `CodeIndexer` by restoring a persisted HNSW
-//! snapshot and chunk corpus from disk.
-//!
-//! Why (issue #85): both `POST /indexes` and `restore_indexes` need the same
-//! logic — construct the indexer, wire the embedder, load HNSW + chunks, fall
-//! back on any soft failure. Centralising prevents drift between call sites.
-//! What: `build_indexer_from_entry` returns `Result<CodeIndexer, anyhow::Error>`.
-//! Corrupt/missing snapshots fall back to an empty store (logged at WARN/INFO).
-//! OOM allocating the HNSW store surfaces as `Err` so callers can skip the
-//! index cleanly instead of panicking (closes #954).
-//! Test: integration tests in `tests/integration_tests.rs`.
-
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-
-use anyhow::Result;
-use trusty_common::migrations::{
-    file_stamp::{read_version_from_file, write_version_to_file},
-    MigrationRunner, SchemaVersion,
-};
-
-use crate::core::{
-    corpus::CorpusStore,
-    embed::Embedder,
-    indexer::{migrations::JsonCorpusToRedbMigration, CodeIndexer},
-    store::{UsearchStore, VectorStore},
-};
-
-use crate::service::persistence::{self, PersistedIndex};
-
-/// Open a `CorpusStore`, retrying once on `DatabaseAlreadyOpen` (issue #840).
-///
-/// Why: on a fast daemon restart the OS may not have released redb's file lock.
-/// A 50 ms async sleep avoids blocking a tokio worker thread during the retry.
-/// What: calls `CorpusStore::open`; on `DatabaseError::DatabaseAlreadyOpen`
-/// (matched via typed downcast) sleeps 50 ms and retries once. All other
-/// errors surface immediately.
-/// Test: `corpus_recovery::tests::database_already_open_variant_is_stable`
-/// (pinning) + warm-boot tests in this module.
-async fn open_corpus_with_retry(path: &Path) -> Result<CorpusStore> {
-    match CorpusStore::open(path) {
-        Ok(store) => Ok(store),
-        Err(e) => {
-            // Typed downcast — redb error-message rewording cannot disable retry.
-            let is_already_open = e
-                .downcast_ref::<redb::DatabaseError>()
-                .map(|db_err| matches!(db_err, redb::DatabaseError::DatabaseAlreadyOpen))
-                .unwrap_or(false);
-            if is_already_open {
-                tracing::warn!(
-                    "warm-boot: redb corpus at {} is locked (DatabaseAlreadyOpen) — \
-                     retrying in 50 ms (refs #840)",
-                    path.display()
-                );
-                tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-                CorpusStore::open(path)
-            } else {
-                Err(e)
-            }
-        }
-    }
-}
-
-/// Build a `CodeIndexer` for `index_id`, restoring HNSW + chunks from disk.
-///
-/// Why: see module docs. Returns `Err` only on OOM allocating the fresh store.
-/// What: tries `UsearchStore::load_from`; falls back to fresh on missing/corrupt
-/// snapshot (logged at WARN); attaches embedder + store; rehydrates corpus.
-/// Test: see module docs.
-pub async fn build_indexer_with_persisted_state(
-    index_id: &str,
-    root_path: PathBuf,
-    embedder: &Arc<dyn Embedder>,
-) -> Result<CodeIndexer> {
-    // Build a minimal PersistedIndex so we can use the entry-aware path helpers.
-    // `colocated` defaults to false (legacy global storage) for backward
-    // compatibility — callers that have a full `PersistedIndex` should call
-    // `build_indexer_from_entry` instead.
-    let entry = PersistedIndex {
-        id: index_id.to_string(),
-        root_path: root_path.clone(),
-        ..Default::default()
-    };
-    build_indexer_from_entry(&entry, embedder).await
-}
-
-/// Build a `CodeIndexer` from a `PersistedIndex`, routing storage to colocated
-/// or legacy global paths based on `entry.colocated`.
-///
-/// Why: preserves the `colocated` flag — colocated indexes open storage from
-/// `<root_path>/.trusty-search/`; legacy from the global data dir. Returns
-/// `Err` only on OOM; corrupt/missing snapshots fall back to an empty store.
-/// What: resolves paths via `hnsw_path_for_entry` / `corpus_redb_path_for_entry`,
-/// builds the store (with fallback), attaches the embedder, rehydrates corpus.
-/// Test: `colocated_indexer_builds_from_entry` covers the colocated path;
-/// the existing warm-boot integration tests cover the legacy path.
-pub async fn build_indexer_from_entry(
-    entry: &PersistedIndex,
-    embedder: &Arc<dyn Embedder>,
-) -> Result<CodeIndexer> {
-    let index_id = &entry.id;
-    let root_path = entry.root_path.clone();
-    let dim = embedder.dimension();
-    let (store, hnsw_load_failed): (Arc<dyn VectorStore>, bool) =
-        build_store_for_entry(entry, dim).await?;
-    let mut indexer =
-        CodeIndexer::new(index_id, root_path).with_components(Arc::clone(embedder), store);
-    // Issue #2922: propagate whether a persisted snapshot existed but failed
-    // to load, so warm-boot / lazy-load can fold it into `hnsw_snapshot_ready`
-    // instead of trusting bare file existence.
-    indexer.hnsw_load_failed = hnsw_load_failed;
-    // Issue #313 / #2984 (Phase 0): propagate `skip_kg` onto the indexer
-    // itself — not just the `IndexHandle` built later by the caller — so
-    // `restore_corpus_for_entry` below (via `load_chunks_from_redb` /
-    // `load_chunks_from_disk`) can skip the symbol-graph load/rebuild
-    // entirely instead of paying the heap cost the flag was meant to avoid.
-    indexer.skip_kg = entry.skip_kg;
-
-    // Issue #28/#840/#1158: wire the durable redb corpus store.  Failure is
-    // non-fatal but logged at ERROR (#840) because a missing corpus means the
-    // next reindex cold-starts (Skipped 0).  `open_corpus_with_retry` retries
-    // once on DatabaseAlreadyOpen (stale file lock from a rapid restart).
-    // Issue #1158: set `corpus_open_failed` so the warm-boot stage-classifier
-    // can emit `StageStatus::Failed` instead of the misleading `InProgress`.
-    match persistence::corpus_redb_path_for_entry(entry) {
-        Ok(redb_path) => {
-            let open_result = open_corpus_with_retry(&redb_path).await;
-            match open_result {
-                Ok(corpus) => indexer.set_corpus_store(Arc::new(corpus)),
-                Err(e) => {
-                    tracing::error!(
-                        "warm-boot: FAILED to open redb corpus for '{index_id}' at {} ({e}). \
-                         The durable corpus store is unavailable — the next reindex will be a \
-                         full cold-start (Skipped 0). Check permissions and whether another \
-                         process holds the redb file lock. (refs #840, #1158)",
-                        redb_path.display()
-                    );
-                    // Issue #1158: mark the failure so the stage-classifier
-                    // emits Failed instead of the misleading InProgress.
-                    indexer.corpus_open_failed = true;
-                }
-            }
-        }
-        Err(e) => tracing::warn!("cannot resolve redb corpus path for '{index_id}': {e}"),
-    }
-
-    restore_corpus_for_entry(&mut indexer, entry).await;
-    Ok(indexer)
-}
-
-/// Restore the chunk corpus for `indexer`, preferring the redb store and
-/// falling back to the legacy `chunks.json` snapshot via the shared
-/// `trusty-common::migrations` runner (issues #28, #179).
-///
-/// Why: redb is the new source of truth, but daemons upgraded in place have a
-/// populated `chunks.json` and an empty `index.redb`. The migration kernel
-/// owns the ordering — we always try redb first (the fast path) and only run
-/// the legacy-JSON step when redb is empty *and* the on-disk schema stamp
-/// hasn't already recorded that the migration ran. The stamp is written
-/// after every successful migration step so subsequent boots skip the JSON
-/// probe entirely.
-/// What: tries `load_chunks_from_redb` first; on a populated redb corpus
-/// stamps the schema (so legacy `chunks.json` becomes inert) and returns.
-/// On an empty redb the [`MigrationRunner`] dispatches
-/// [`JsonCorpusToRedbMigration`], which reads the JSON snapshot and seeds
-/// redb. The runner writes the schema stamp after each successful step.
-/// Test: covered by the corpus roundtrip + migration integration tests; the
-/// runner itself is unit-tested in `trusty-common::migrations`.
-async fn restore_corpus_for_entry(indexer: &mut CodeIndexer, entry: &PersistedIndex) {
-    let index_id = &entry.id;
-    // Primary path: redb durable corpus.
-    match indexer.load_chunks_from_redb().await {
-        Ok(n) if n > 0 => {
-            tracing::info!("warm-boot: restored {n} chunks for index '{index_id}' from redb");
-            // Ensure the schema stamp is bumped past the JSON → redb step so
-            // a stray `chunks.json` left over from the legacy build is never
-            // re-read on the next boot.
-            stamp_if_unversioned_for_entry(entry);
-            return;
-        }
-        Ok(_) => {} // empty redb — fall through to the migration runner.
-        Err(e) => tracing::warn!(
-            "warm-boot: redb corpus load failed for '{index_id}' ({e}) — \
-             trying registered migrations"
-        ),
-    }
-
-    // Migration runner path (issue #179): dispatches the legacy JSON →
-    // redb migration when the on-disk schema stamp says it hasn't yet run.
-    run_migrations_for_entry(indexer, entry);
-}
-
-/// Dispatch the trusty-search migration runner for one index.
-///
-/// Why: lifted into its own function so the redb-empty branch in
-/// [`restore_corpus_for_entry`] and any future "force re-migrate" admin command
-/// can share one entry point. Keeps the runner's stamp file path resolution
-/// and error logging in one place. Uses `schema_version_path_for_entry` so the
-/// stamp lands in the right location for both colocated and legacy indexes.
-/// What: reads the current stamp via `read_version_from_file`, instantiates
-/// the runner with [`JsonCorpusToRedbMigration`], and runs it against the
-/// indexer. Failures are logged but never propagated — a missing
-/// `chunks.json` is the genuine first-boot case and yields an empty corpus.
-/// Test: covered by the existing migration integration tests.
-fn run_migrations_for_entry(indexer: &mut CodeIndexer, entry: &PersistedIndex) {
-    let index_id = &entry.id;
-    let stamp_path = match persistence::schema_version_path_for_entry(entry) {
-        Ok(p) => p,
-        Err(e) => {
-            tracing::warn!("cannot resolve schema version path for '{index_id}': {e}");
-            return;
-        }
-    };
-    let current = match read_version_from_file(&stamp_path) {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::warn!(
-                "warm-boot: failed to read schema stamp at {} ({e}) — \
-                 treating as UNVERSIONED",
-                stamp_path.display()
-            );
-            SchemaVersion::UNVERSIONED
-        }
-    };
-    let runner = MigrationRunner::new(vec![Box::new(JsonCorpusToRedbMigration)]);
-    if let Err(e) = runner.run(indexer, current, |v| write_version_to_file(&stamp_path, v)) {
-        tracing::warn!(
-            "warm-boot: migration runner failed for '{index_id}' ({e}) — \
-             starting with whatever state was restored"
-        );
-    }
-}
-
-/// Stamp the schema as fully migrated if the on-disk stamp is currently
-/// UNVERSIONED.
-///
-/// Why: the redb-populated branch in [`restore_corpus_for_entry`] short-circuits
-/// without invoking the runner. We still want the stamp to advance so a
-/// future migration with a higher `from_version` runs cleanly, and so a
-/// stray legacy `chunks.json` is never reprocessed. Writing only when the
-/// stamp is missing/UNVERSIONED preserves any version that a previous
-/// runner write recorded.
-/// What: best-effort — read the stamp, and if it is UNVERSIONED, write the
-/// current `TRUSTY_SEARCH_SCHEMA_TARGET`. A failure is logged at WARN.
-/// Test: covered indirectly via the corpus roundtrip integration tests.
-fn stamp_if_unversioned_for_entry(entry: &PersistedIndex) {
-    use crate::core::indexer::migrations::TRUSTY_SEARCH_SCHEMA_TARGET;
-
-    let index_id = &entry.id;
-    let stamp_path = match persistence::schema_version_path_for_entry(entry) {
-        Ok(p) => p,
-        Err(e) => {
-            tracing::warn!("cannot resolve schema version path for '{index_id}': {e}");
-            return;
-        }
-    };
-    let current = read_version_from_file(&stamp_path).unwrap_or(SchemaVersion::UNVERSIONED);
-    if current >= TRUSTY_SEARCH_SCHEMA_TARGET {
-        return;
-    }
-    if let Err(e) = write_version_to_file(&stamp_path, TRUSTY_SEARCH_SCHEMA_TARGET) {
-        tracing::warn!(
-            "warm-boot: failed to bump schema stamp for '{index_id}' at {} ({e})",
-            stamp_path.display()
-        );
-    }
-}
-
-/// Try to load the HNSW snapshot for `entry`. On soft failure (missing, corrupt,
-/// dim mismatch) falls back to a fresh empty store. Returns `Err` only on OOM.
-///
-/// Why: uses `hnsw_path_for_entry` so colocated indexes read from the project's
-/// `.trusty-search/hnsw.usearch`; propagates OOM as `Err` (closes #954).
-/// What: resolves path, checks for snapshot, loads (or falls back), returns
-/// `(store, load_failed)`. `load_failed` is `true` only when a snapshot file
-/// existed but could not be restored (issue #2922) — never for the ordinary
-/// "no snapshot yet" first-boot case — so callers can distinguish "never
-/// indexed" from "corrupt/truncated on-disk state" when deriving
-/// `hnsw_snapshot_ready` for `/health`.
-/// Test: warm-boot integration tests (legacy) + colocated integration tests;
-/// `tests::build_store_for_entry_flags_load_failure_on_corrupt_snapshot`.
-async fn build_store_for_entry(
-    entry: &PersistedIndex,
-    dim: usize,
-) -> Result<(Arc<dyn VectorStore>, bool)> {
-    let index_id = &entry.id;
-    let path = match persistence::hnsw_path_for_entry(entry) {
-        Ok(p) => p,
-        Err(e) => {
-            tracing::warn!("cannot resolve hnsw path for '{index_id}': {e}");
-            return Ok((fresh_store(dim)?, false));
-        }
-    };
-
-    if persistence::has_persisted_hnsw(&path) {
-        match UsearchStore::load_from(&path).await {
-            Ok(Some(store)) => {
-                if store.dim() == dim {
-                    tracing::info!(
-                        "warm-boot: restored HNSW snapshot for '{}' from {}",
-                        index_id,
-                        path.display()
-                    );
-                    return Ok((Arc::new(store), false));
-                }
-                tracing::warn!(
-                    "warm-boot: hnsw snapshot for '{}' has dim {} but embedder is {} — starting fresh",
-                    index_id,
-                    store.dim(),
-                    dim
-                );
-            }
-            Ok(None) => {
-                // Sidecar missing/corrupt, or the #2922 zero-vector-vs-populated
-                // guard tripped — fall back to fresh.
-                tracing::warn!(
-                    "warm-boot: hnsw snapshot at {} could not be loaded — starting fresh \
-                     (semantic search will report not-ready for '{}' until reindexed, issue #2922)",
-                    path.display(),
-                    index_id,
-                );
-            }
-            Err(e) => {
-                tracing::warn!(
-                    "warm-boot: error loading hnsw snapshot at {}: {e} — starting fresh \
-                     (semantic search will report not-ready for '{}' until reindexed, issue #2922)",
-                    path.display(),
-                    index_id,
-                );
-            }
-        }
-        // A snapshot file existed but we fell through to the fresh-store
-        // fallback above — issue #2922: the caller must NOT treat this index
-        // as having a ready semantic snapshot.
-        return Ok((fresh_store(dim)?, true));
-    }
-    Ok((fresh_store(dim)?, false))
-}
-
-/// Allocate a fresh empty `UsearchStore`. Returns `Err` on OOM instead of
-/// panicking so callers can skip the index with a diagnostic log (#954).
-///
-/// Why: the HNSW allocator can fail on memory-constrained hosts at warm-boot;
-/// propagating the error lets the daemon keep serving remaining indexes rather
-/// than aborting. Systemd `Restart=on-failure` handles persistent OOM.
-/// What: calls `UsearchStore::new(dim)`; logs ERROR and returns `Err` on failure.
-/// Test: normal allocation exercised by all persistence_loader tests; OOM path
-/// is not reproducible in unit tests (requires near-full RAM).
-fn fresh_store(dim: usize) -> Result<Arc<dyn VectorStore>> {
-    UsearchStore::new(dim)
-        .map(|s| Arc::new(s) as Arc<dyn VectorStore>)
-        .map_err(|e| {
-            tracing::error!(
-                "failed to allocate UsearchStore (dim={dim}): {e} — \
-                 index will be skipped (OOM, #954)"
-            );
-            anyhow::anyhow!("usearch alloc failure (OOM, dim={dim}): {e}")
-        })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::chunker::{ChunkType, RawChunk};
-    use crate::service::colocated_storage;
-    use crate::service::persistence::PersistedIndex;
-    use tempfile::tempdir;
-    use trusty_common::embedder::MockEmbedder;
-
-    // ── Helper ────────────────────────────────────────────────────────────────
-
-    fn mock_embedder() -> Arc<dyn crate::core::embed::Embedder> {
-        Arc::new(MockEmbedder::new(8))
-    }
-
-    fn minimal_raw_chunk(id: &str) -> RawChunk {
-        RawChunk {
-            id: id.to_string(),
-            file: "src/lib.rs".to_string(),
-            start_line: 1,
-            end_line: 5,
-            content: "fn hello() {}".to_string(),
-            function_name: None,
-            language: Some("rust".to_string()),
-            chunk_type: ChunkType::Code,
-            calls: Vec::new(),
-            inherits_from: Vec::new(),
-            chunk_depth: 0,
-            parent_chunk_id: None,
-            child_chunk_ids: Vec::new(),
-            nlp_keywords: Vec::new(),
-            nlp_code_refs: Vec::new(),
-            virtual_terms: Vec::new(),
-        }
-    }
-
-    // ── Issue #483 regression ─────────────────────────────────────────────────
-
-    /// Why: guards the writer/loader path divergence (#483) where
-    /// `create_index_handler` used `colocated: false`, routing the corpus to
-    /// app-data, while warm-boot used `colocated: true`, opening an empty
-    /// `.trusty-search/` dir.
-    /// What: builds a colocated indexer, writes a chunk, drops the handle
-    /// (releasing the redb file lock), then reloads via `build_indexer_from_entry`
-    /// and asserts the chunk count is > 0.
-    /// Test: this IS the test; exercises the exact call path used by the fixed
-    /// `create_index_handler`.
-    #[tokio::test]
-    async fn colocated_create_handler_path_survives_simulated_reload() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        let embedder = mock_embedder();
-
-        // --- Phase 1: simulate what the fixed create_index_handler does ---
-        let entry = PersistedIndex {
-            id: "test-idx-483".to_string(),
-            root_path: root.clone(),
-            colocated: true,
-            ..Default::default()
-        };
-        // build_indexer_from_entry (the fixed call path): colocated_storage_dir
-        // is invoked via corpus_redb_path_for_entry → colocated_redb_path →
-        // colocated_storage_dir, creating `.trusty-search/` on disk.
-        let indexer = build_indexer_from_entry(&entry, &embedder).await.unwrap();
-
-        // The corpus store must be wired and pointing into `.trusty-search/`.
-        assert!(
-            indexer.has_corpus_store(),
-            "#483: indexer must have a corpus store after build_indexer_from_entry with colocated=true"
-        );
-        // `.trusty-search/` must exist so the write-path probes see it.
-        assert!(
-            colocated_storage::has_colocated_storage(&root),
-            "#483: .trusty-search/ must exist after build_indexer_from_entry with colocated=true"
-        );
-
-        // Write a chunk to the wired corpus so the reload can verify it.
-        // Scope the write so the corpus Arc is dropped before we re-open
-        // the same redb file for the simulated restart (redb uses file locking
-        // and rejects a second concurrent open of the same database file).
-        {
-            let corpus = indexer.corpus_store().expect("corpus store must be set");
-            corpus
-                .upsert_chunks(&[minimal_raw_chunk("src/lib.rs:1:5")])
-                .expect("upsert must succeed");
-            // corpus Arc dropped here.
-        }
-        // Drop the whole indexer (and its internal corpus Arc) before reopening.
-        drop(indexer);
-
-        // --- Phase 2: simulate daemon restart — reload from the same entry ---
-        // This is what `restore_indexes` does on startup.
-        let reloaded = build_indexer_from_entry(&entry, &embedder).await.unwrap();
-        assert!(
-            reloaded.has_corpus_store(),
-            "#483: reloaded indexer must have a corpus store"
-        );
-        let chunk_count = reloaded
-            .corpus_store()
-            .expect("corpus store must be set")
-            .chunk_count()
-            .expect("chunk_count must succeed");
-        assert!(
-            chunk_count > 0,
-            "#483: reloaded index must contain the written chunk (got {chunk_count}); \
-             writer/loader paths must agree on the colocated location"
-        );
-    }
-
-    // ── Issue #2922 regression ────────────────────────────────────────────────
-
-    /// Why: this is the acceptance test for the "health reports semantic
-    /// ready while silently BM25-only" half of issue #2922. Before the fix,
-    /// `hnsw_snapshot_ready` (and therefore `hnsw_load_failed`) had no way to
-    /// tell "a file exists" apart from "a file exists AND loaded correctly" —
-    /// a truncated `.usearch` snapshot still passed `has_persisted_hnsw`.
-    /// What: saves a real HNSW snapshot with one vector via `UsearchStore`,
-    /// truncates the on-disk `.usearch` file to simulate the exact corruption
-    /// reported in #2922 (a timed-out shutdown flush cutting a save short),
-    /// then calls `build_indexer_from_entry` and asserts `hnsw_load_failed`
-    /// is `true` and the restored store is empty (never a silent partial
-    /// state).
-    /// Test: this IS the test.
-    #[tokio::test]
-    async fn build_store_for_entry_flags_load_failure_on_corrupt_snapshot() {
-        use crate::core::store::{UsearchStore, VectorStore};
-
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        let embedder = mock_embedder();
-        let dim = embedder.dimension();
-
-        let entry = PersistedIndex {
-            id: "test-idx-2922".to_string(),
-            root_path: root.clone(),
-            colocated: true,
-            ..Default::default()
-        };
-        let hnsw_path = persistence::hnsw_path_for_entry(&entry).unwrap();
-
-        // Write a real, valid snapshot containing one vector.
-        let store = UsearchStore::new(dim).unwrap();
-        store
-            .upsert("chunk-1", vec![0.1; dim])
-            .await
-            .expect("upsert must succeed");
-        store.save(&hnsw_path).await.expect("save must succeed");
-        assert!(hnsw_path.exists(), "precondition: snapshot must exist");
-
-        // Simulate the #2922 corruption: a shutdown flush cut off mid-write,
-        // leaving a truncated (but still `path.exists()`-true) file.
-        std::fs::write(&hnsw_path, b"truncated").expect("truncate hnsw.usearch");
-
-        let (loaded_store, load_failed): (Arc<dyn VectorStore>, bool) =
-            build_store_for_entry(&entry, dim).await.unwrap();
-        assert!(
-            load_failed,
-            "#2922: a snapshot file that exists but fails to load must set load_failed=true"
-        );
-        assert_eq!(
-            loaded_store.len().await.unwrap(),
-            0,
-            "#2922: a corrupt snapshot must never silently restore partial/wrong data"
-        );
-
-        // End-to-end: `build_indexer_from_entry` must propagate the flag onto
-        // the indexer so the warm-boot / lazy-load call sites can fold it
-        // into `hnsw_snapshot_ready`.
-        let indexer = build_indexer_from_entry(&entry, &embedder).await.unwrap();
-        assert!(
-            indexer.hnsw_load_failed,
-            "#2922: CodeIndexer::hnsw_load_failed must be propagated from build_store_for_entry"
-        );
-    }
-
-    /// Why: the ordinary first-boot case (no snapshot has ever been written)
-    /// must NOT be flagged as a load failure — only a snapshot that existed
-    /// and couldn't be restored is a failure (issue #2922).
-    /// What: builds an indexer for an entry with no on-disk HNSW snapshot at
-    /// all and asserts `hnsw_load_failed` stays `false`.
-    /// Test: this IS the test.
-    #[tokio::test]
-    async fn build_store_for_entry_does_not_flag_first_boot_as_failure() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        let embedder = mock_embedder();
-
-        let entry = PersistedIndex {
-            id: "test-idx-2922-firstboot".to_string(),
-            root_path: root.clone(),
-            colocated: true,
-            ..Default::default()
-        };
-        let indexer = build_indexer_from_entry(&entry, &embedder).await.unwrap();
-        assert!(
-            !indexer.hnsw_load_failed,
-            "#2922: a never-indexed entry (no snapshot file at all) must not be \
-             flagged as a load failure"
-        );
-    }
-
-    // ── Issue #485 regression ─────────────────────────────────────────────────
-
-    /// Why: guards the "cannot write schema_version: no durable corpus" error
-    /// (#485) that occurred when the corpus store was not wired for a colocated
-    /// entry.
-    /// What: builds an indexer via the colocated entry path and asserts both
-    /// `has_corpus_store` and that the corpus path is inside the project root.
-    /// Test: this IS the test.
-    #[tokio::test]
-    async fn colocated_create_path_wires_corpus_store_for_schema_version() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        let embedder = mock_embedder();
-
-        let entry = PersistedIndex {
-            id: "test-idx-485".to_string(),
-            root_path: root.clone(),
-            colocated: true,
-            ..Default::default()
-        };
-        let indexer = build_indexer_from_entry(&entry, &embedder).await.unwrap();
-
-        // Corpus store must be present — this is the prerequisite that prevents
-        // the "cannot write schema_version: no durable corpus" error (#485).
-        assert!(
-            indexer.has_corpus_store(),
-            "#485: indexer built via colocated create path must have corpus store; \
-             without it write_schema_version returns 'no durable corpus'"
-        );
-
-        // Also confirm the store is backed by the colocated path, not app-data.
-        if let Some(corpus) = indexer.corpus_store() {
-            let corpus_path = corpus.path().to_path_buf();
-            assert!(
-                corpus_path.starts_with(&root),
-                "#485: corpus store must be inside the project root (colocated); \
-                 got {corpus_path:?}"
-            );
-        }
-    }
-
-    // ── Guard: legacy colocated=false path is unchanged ───────────────────────
-
-    /// Why: the fix must not break legacy (`colocated: false`) indexes.
-    /// What: builds with `colocated: false`; asserts no panic even when no
-    /// app-data corpus exists.
-    /// Test: this IS the test.
-    #[tokio::test]
-    async fn legacy_non_colocated_path_does_not_panic() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        let embedder = mock_embedder();
-
-        let entry = PersistedIndex {
-            id: "test-idx-legacy".to_string(),
-            root_path: root.clone(),
-            colocated: false,
-            ..Default::default()
-        };
-        // Must not fail even when no app-data corpus exists.
-        build_indexer_from_entry(&entry, &embedder).await.unwrap();
-        // Key invariant: no error. Corpus store presence is not asserted here:
-        // app-data path may or may not resolve in a test environment.
-    }
-
-    // ── Issue #2305 regression: shared-root redb double-open ───────────────────
-
-    /// Why (issue #2305 — the root cause behind #1870): two colocated
-    /// registrations of the SAME `root_path` resolve to one
-    /// `<root>/.trusty-search/index.redb`. redb is a single-open database, so
-    /// warm-booting both opens the file twice and the second fails with
-    /// `DatabaseAlreadyOpen` (surfaced as `corpus_open_failed`) — the 50 ms
-    /// retry can never clear an in-process holder, so it recurs every restart.
-    /// This test proves the hazard is real (a second, concurrently-held open of
-    /// the shared redb fails) and that `dedup_entries_by_corpus_path` collapses
-    /// the pair to exactly one entry so the warm-boot open path opens once.
-    /// What: opens `apex`, then — while that handle is still held — opens `mpm`
-    /// at the same root and asserts `mpm` reports `corpus_open_failed`; then runs
-    /// the pair through the dedup and asserts it collapses to a single survivor.
-    /// Only one *successful* corpus open is performed (the failing second open
-    /// returns immediately) so the test stays cheap. Data-survival across a
-    /// reopen is already covered by
-    /// `colocated_create_handler_path_survives_simulated_reload`, and survivor
-    /// selection by `dedup_by_corpus_path_keeps_most_recent`.
-    /// Test: this IS the test.
-    #[tokio::test]
-    async fn warm_boot_shared_root_double_open_is_prevented_by_dedup() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        let embedder = mock_embedder();
-
-        let apex = PersistedIndex {
-            id: "apex".to_string(),
-            root_path: root.clone(),
-            colocated: true,
-            last_indexed_unix: Some(100),
-            ..Default::default()
-        };
-        let mpm = PersistedIndex {
-            id: "trusty-mpm".to_string(),
-            root_path: root.clone(),
-            colocated: true,
-            last_indexed_unix: Some(50),
-            ..Default::default()
-        };
-
-        // Prove the hazard: the first open succeeds and holds the corpus; a
-        // second open of the same redb file — the exact shape of the sequential
-        // warm-boot restore over two shared-root slots — fails with
-        // DatabaseAlreadyOpen (surfaced as corpus_open_failed, the #1870 symptom).
-        let first = build_indexer_from_entry(&apex, &embedder).await.unwrap();
-        let second = build_indexer_from_entry(&mpm, &embedder).await.unwrap();
-        assert!(
-            first.has_corpus_store() && !first.corpus_open_failed,
-            "the first open of the shared redb must succeed"
-        );
-        assert!(
-            second.corpus_open_failed,
-            "#2305: a concurrent second open of the shared redb must fail \
-             (DatabaseAlreadyOpen) — this is the bug the dedup prevents"
-        );
-        drop(first);
-        drop(second);
-
-        // Apply the fix: dedup collapses the two shared-root entries to one, so
-        // the warm-boot restore opens the shared redb exactly once → zero
-        // DatabaseAlreadyOpen. (The survivor is the most recent, `apex`.)
-        let survivors = crate::service::warm_boot::dedup_entries_by_corpus_path(vec![apex, mpm]);
-        assert_eq!(
-            survivors.len(),
-            1,
-            "#2305: two entries sharing one redb corpus must collapse to one"
-        );
-        assert_eq!(
-            survivors[0].id, "apex",
-            "the most-recently-active shared-root entry must be the survivor"
-        );
-    }
-
-    /// Why: the fix must not merge indexes that live in different redb files.
-    /// Two colocated indexes at distinct roots must open independent corpora
-    /// simultaneously with zero `DatabaseAlreadyOpen`, and dedup must keep both.
-    /// What: builds two colocated indexers at distinct roots while both handles
-    /// are held; asserts both are healthy and dedup returns both entries.
-    /// Test: this IS the test.
-    #[tokio::test]
-    async fn distinct_roots_open_independent_corpora() {
-        let a = tempdir().unwrap();
-        let b = tempdir().unwrap();
-        let embedder = mock_embedder();
-
-        let ea = PersistedIndex {
-            id: "a".to_string(),
-            root_path: a.path().to_path_buf(),
-            colocated: true,
-            ..Default::default()
-        };
-        let eb = PersistedIndex {
-            id: "b".to_string(),
-            root_path: b.path().to_path_buf(),
-            colocated: true,
-            ..Default::default()
-        };
-
-        // Both held open at once — distinct redb files, so no double-open.
-        let ia = build_indexer_from_entry(&ea, &embedder).await.unwrap();
-        let ib = build_indexer_from_entry(&eb, &embedder).await.unwrap();
-        assert!(
-            ia.has_corpus_store() && !ia.corpus_open_failed,
-            "distinct-root corpus a must open cleanly"
-        );
-        assert!(
-            ib.has_corpus_store() && !ib.corpus_open_failed,
-            "distinct-root corpus b must open cleanly"
-        );
-
-        // Dedup must leave distinct roots untouched.
-        let kept = crate::service::warm_boot::dedup_entries_by_corpus_path(vec![ea, eb]);
-        assert_eq!(
-            kept.len(),
-            2,
-            "distinct roots resolve to distinct redb files and must both survive dedup"
-        );
-    }
-}
